[#2697 #2706] Rename SearchMarkerUtils.js to LocationMarkerUtils.js
