+++ lib/Models/LocationMarkerUtils.js
@@ -67,6 +67,6 @@ export function removeMarker(terria, viewState) {
 }
 
 /**
- * The name given to the data source created by {@link #addMarker}. 
+ * The name given to the data source created by {@link #addMarker}.
  */
 export { SEARCH_MARKER_DATA_SOURCE_NAME };
