+++ lib/ReactViews/Mobile/MobileHeader.jsx
@@ -9,7 +9,7 @@ import Styles from './mobile-header.scss';
 import Icon from "../Icon.jsx";
 import MobileMenu from './MobileMenu';
 import classNames from 'classnames';
-import { removeMarker } from '../../Models/SearchMarkerUtils';
+import { removeMarker } from '../../Models/LocationMarkerUtils';
 
 const MobileHeader = createReactClass({
     displayName: 'MobileHeader',
