+++ lib/ReactViews/Search/SidebarSearch.jsx
@@ -7,7 +7,7 @@ import BadgeBar from '../BadgeBar.jsx';
 import Styles from './sidebar-search.scss';
 import LocationSearchResults from './LocationSearchResults.jsx';
 
-import {addMarker} from '../../Models/SearchMarkerUtils';
+import {addMarker} from '../../Models/LocationMarkerUtils';
 
 // Handle any of the three kinds of search based on the props
 const SidebarSearch = createReactClass({
